refactor(app-distribution): remove deprecated namespaced API
