+++ packages/app-distribution/lib/types/internal.ts
@@ -1,7 +1,5 @@
 import type { AppDistribution, AppDistributionRelease } from './app-distribution';
 
-export type AppDistributionModularDeprecationArg = string;
-
 export interface RNFBAppDistributionModule {
   isTesterSignedIn(): Promise<boolean>;
   signInTester(): Promise<void>;
@@ -16,12 +14,9 @@ declare module '@react-native-firebase/app/dist/module/internal/NativeModules' {
 }
 
 export interface AppDistributionInternal extends AppDistribution {
+  isTesterSignedIn(): Promise<boolean>;
+  signInTester(): Promise<void>;
+  checkForUpdate(): Promise<AppDistributionRelease>;
+  signOutTester(): Promise<void>;
   readonly native: RNFBAppDistributionModule;
 }
-
-export type AppDistributionWithDeprecationArg = AppDistributionInternal & {
-  isTesterSignedIn(_depArg: AppDistributionModularDeprecationArg): Promise<boolean>;
-  signInTester(_depArg: AppDistributionModularDeprecationArg): Promise<void>;
-  checkForUpdate(_depArg: AppDistributionModularDeprecationArg): Promise<AppDistributionRelease>;
-  signOutTester(_depArg: AppDistributionModularDeprecationArg): Promise<void>;
-};
